requirements.txt
annotated-types==0.7.0
anyio==4.10.0
black==25.1.0
certifi==2025.8.3
click==8.2.1
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.3
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-multipart==0.0.20
ruff==0.12.10
setuptools==70.2.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.3
torch==2.8.0+cpu
torchvision==0.23.0+cpu
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
